requirements.txt
backports-abc==0.5
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
bleach==3.3.1
certifi==2020.6.20
cffi==1.14.0
cloudpickle==1.2.2
configparser==4.0.2
cycler==0.10.0
cytoolz==0.10.1
dask==1.2.2
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
enum34==1.1.6
functools32==3.2.3.post2
future==0.18.2
futures==3.3.0
h5py==2.8.0
imageio==2.6.1
ipaddress==1.0.23
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.6.0
Jinja2==2.11.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==5.2.0
jupyter-core==4.6.1
kiwisolver==1.1.0
leveldb==0.20
linecache2==1.0.0
MarkupSafe==1.1.1
matplotlib==2.2.3
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.2
notebook==5.7.10
numpy==1.16.6
olefile==0.46
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
packaging==20.9
palettable==3.3.0
pandas==0.22.0
pandocfilters==1.4.2
pathlib2==2.3.5
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.11.0
prompt-toolkit==1.0.15
protobuf==3.11.2
ptyprocess==0.7.0
pycparser==2.20
pydotplus==2.0.2
Pygments==2.5.2
pyparsing==2.4.7
pypng==0.0.20
python-dateutil==2.8.1
python-gflags==3.1.2
pytz==2021.1
PyWavelets==1.0.3
PyYAML==5.2
pyzmq==18.1.0
qtconsole==4.7.7
QtPy==1.10.0
scandir==1.10.0
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.8.0
simplegeneric==0.8.1
singledispatch==0.0.0
six==1.15.0
statsmodels==0.10.1
subprocess32==3.5.4
svgwrite==1.3.1
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
torch==1.3.1
torchvision==0.4.2
tornado==5.1.1
traceback2==1.4.0
traitlets==4.3.3
typing==3.7.4.1
unittest2==1.1.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
XlsxWriter==2.0.0
xlwt==1.2.0
